backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = {extras = ["all"], version = "^0.87.0"}
uvicorn = "^0.20.0"
sse-starlette = "^1.2.1"
asyncio = "^3.4.3"
confluent-kafka = "^1.9.2"
pydantic = "^1.10.2"
click = "^8.1.3"
python-dotenv = "^0.21.0"


[tool.poetry.group.dev.dependencies]
sseclient-py = "^1.7.2"
requests = "^2.28.1"
isort = "^5.11.2"
black = "^22.12.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
